requirements.txt
black==24.4.2
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.7
pyotp==2.9.0
python-dotenv==1.0.0
twilio==9.0.5
